requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
phonenumberslite==8.13.4
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
zipp==3.11.0
